+++ example/lib/main.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: public_member_api_docs
+
 import 'dart:io';
 
 import 'package:flutter/material.dart';
@@ -7,9 +9,9 @@ import 'package:flutter_uploader_example/upload_screen.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 
-const String title = "FileUpload Sample app";
+const String title = 'FileUpload Sample app';
 final Uri uploadURL = Uri.parse(
-  "https://us-central1-flutteruploadertest.cloudfunctions.net/upload",
+  'https://us-central1-flutteruploadertest.cloudfunctions.net/upload',
 );
 
 final azureConnectionString = '';
@@ -21,10 +23,9 @@ void backgroundHandler() {
   WidgetsFlutterBinding.ensureInitialized();
 
   // Notice these instances belong to a forked isolate.
-  FlutterUploader uploader = FlutterUploader();
+  var uploader = FlutterUploader();
 
-  FlutterLocalNotificationsPlugin notifications =
-      FlutterLocalNotificationsPlugin();
+  var notifications = FlutterLocalNotificationsPlugin();
 
   // Only show notifications for unprocessed uploads.
   SharedPreferences.getInstance().then((preferences) {
@@ -46,20 +47,20 @@ void backgroundHandler() {
           'FlutterUploader Example',
           'Upload in Progress',
           NotificationDetails(
-            AndroidNotificationDetails(
+            android: AndroidNotificationDetails(
               'FlutterUploader.Example',
               'FlutterUploader',
               'Installed when you activate the Flutter Uploader Example',
               progress: progress.progress,
               icon: 'ic_upload',
               enableVibration: false,
-              importance: Importance.Low,
+              importance: Importance.low,
               showProgress: true,
               onlyAlertOnce: true,
               maxProgress: 100,
               channelShowBadge: false,
             ),
-            IOSNotificationDetails(),
+            iOS: IOSNotificationDetails(),
           ),
         );
       });
@@ -96,17 +97,17 @@ void backgroundHandler() {
         'FlutterUploader Example',
         title,
         NotificationDetails(
-          AndroidNotificationDetails(
+          android: AndroidNotificationDetails(
             'FlutterUploader.Example',
             'FlutterUploader',
             'Installed when you activate the Flutter Uploader Example',
             icon: 'ic_upload',
             enableVibration: result.status == UploadTaskStatus.complete,
             importance: result.status == UploadTaskStatus.failed
-                ? Importance.High
-                : Importance.Min,
+                ? Importance.high
+                : Importance.min,
           ),
-          IOSNotificationDetails(presentAlert: true),
+          iOS: IOSNotificationDetails(presentAlert: true),
         ),
       )
           .catchError((e, stack) {
@@ -123,6 +124,7 @@ class App extends StatefulWidget {
 
   const App({Key key, this.child}) : super(key: key);
 
+  @override
   _AppState createState() => _AppState();
 }
 
@@ -135,8 +137,7 @@ class _AppState extends State<App> {
 
     _uploader.setBackgroundHandler(backgroundHandler);
 
-    FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
-        FlutterLocalNotificationsPlugin();
+    var flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin();
     var initializationSettingsAndroid =
         AndroidInitializationSettings('ic_upload');
     var initializationSettingsIOS = IOSInitializationSettings(
@@ -147,7 +148,9 @@ class _AppState extends State<App> {
           (int id, String title, String body, String payload) async {},
     );
     var initializationSettings = InitializationSettings(
-        initializationSettingsAndroid, initializationSettingsIOS);
+      android: initializationSettingsAndroid,
+      iOS: initializationSettingsIOS,
+    );
     flutterLocalNotificationsPlugin.initialize(
       initializationSettings,
       onSelectNotification: (payload) async {},