+++ example/test_driver/flutter_uploader_test.dart
@@ -16,17 +16,19 @@ void main() {
   E2EWidgetsFlutterBinding.ensureInitialized();
 
   FlutterUploader uploader;
-  List<String> tempFilePaths = [];
+  var tempFilePaths = <String>[];
 
   setUp(() {
     uploader = FlutterUploader();
   });
 
   tearDownAll(() {
-    for (String path in tempFilePaths) {
+    for (var path in tempFilePaths) {
       try {
         File(path).deleteSync();
-      } catch (e) {}
+      } catch (e) {
+        // ignored
+      }
     }
     tempFilePaths.clear();
   });
@@ -45,8 +47,8 @@ void main() {
   group('multipart/form-data uploads', () {
     final url = baseUrl;
 
-    testWidgets("single file", (WidgetTester tester) async {
-      var fileItem = FileItem(path: await _tmpFile(), field: "file");
+    testWidgets('single file', (WidgetTester tester) async {
+      var fileItem = FileItem(path: await _tmpFile(), field: 'file');
 
       final taskId = await uploader.enqueue(
         MultipartFormDataUpload(url: url.toString(), files: [fileItem]),
@@ -64,7 +66,7 @@ void main() {
     });
 
     testWidgets("can overwrite 'Accept' header", (WidgetTester tester) async {
-      var fileItem = FileItem(path: await _tmpFile(), field: "file");
+      var fileItem = FileItem(path: await _tmpFile(), field: 'file');
 
       final taskId = await uploader.enqueue(MultipartFormDataUpload(
         url: url.toString(),
@@ -78,11 +80,11 @@ void main() {
           'application/json, charset=utf-8');
     });
 
-    testWidgets("multiple files", (WidgetTester tester) async {
+    testWidgets('multiple files', (WidgetTester tester) async {
       final taskId = await uploader.enqueue(
         MultipartFormDataUpload(url: url.toString(), files: [
-          FileItem(path: await _tmpFile(256), field: "file1"),
-          FileItem(path: await _tmpFile(257), field: "file2"),
+          FileItem(path: await _tmpFile(256), field: 'file1'),
+          FileItem(path: await _tmpFile(257), field: 'file2'),
         ]),
       );
 
@@ -96,8 +98,8 @@ void main() {
       expect(res.status, UploadTaskStatus.complete);
     });
 
-    testWidgets("handles 201 empty body", (WidgetTester tester) async {
-      var fileItem = FileItem(path: await _tmpFile(), field: "file");
+    testWidgets('handles 201 empty body', (WidgetTester tester) async {
+      var fileItem = FileItem(path: await _tmpFile(), field: 'file');
 
       final taskId = await uploader.enqueue(
         MultipartFormDataUpload(
@@ -117,8 +119,8 @@ void main() {
       expect(res.status, UploadTaskStatus.complete);
     });
 
-    testWidgets("forwards errors", (WidgetTester tester) async {
-      var fileItem = FileItem(path: await _tmpFile(), field: "file");
+    testWidgets('forwards errors', (WidgetTester tester) async {
+      var fileItem = FileItem(path: await _tmpFile(), field: 'file');
 
       final taskId = await uploader.enqueue(
         MultipartFormDataUpload(
@@ -139,7 +141,7 @@ void main() {
   group('binary uploads', () {
     final url = baseUrl.replace(path: baseUrl.path + 'Binary');
 
-    testWidgets("single file", (WidgetTester tester) async {
+    testWidgets('single file', (WidgetTester tester) async {
       final taskId = await uploader.enqueue(
         RawUpload(
           url: url.toString(),
@@ -170,7 +172,7 @@ void main() {
 
       expect(json['headers']['accept'], 'application/json, charset=utf-8');
     });
-    testWidgets("fowards errors", (WidgetTester tester) async {
+    testWidgets('fowards errors', (WidgetTester tester) async {
       final taskId = await uploader.enqueue(
         RawUpload(
           url: