+++ lib/flutter_uploader.dart
@@ -13,11 +13,7 @@ part 'src/flutter_uploader.dart';
 
 part 'src/upload.dart';
 
-part 'src/upload_exception.dart';
-
 part 'src/upload_method.dart';
-
-part 'src/upload_task.dart';
 
 part 'src/upload_task_progress.dart';